+++ src/routes/index.rs
@@ -24,7 +24,7 @@ pub async fn upload<'a>(
 ) -> RatelimitedRouteResponse<Json<FileData>> {
     let mut ratelimiter = Ratelimiter::new("attachments", "attachments", ip, conf.inner());
     ratelimiter
-        .process_ratelimit(upload.file.len() as u128, &mut cache)
+        .process_ratelimit(upload.file.len(), &mut cache)
         .await?;
     let upload = upload.into_inner();
     let file = File::create(
